create ui base for event selection
