--- resources/js/components/Main.js
@@ -1,100 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import Typography from '@mui/material/Typography';
-import { Button, Card, CardActionArea, CardActions, CardContent, CardMedia, Grid } from '@mui/material';
-
-function Main() {
-    return (
-        <container className='flex justify-center'>
-            <div className='items-center justify-center border h-4/6 w-2/3 my-10 p-5'>
-                <Typography variant='h5' align='center'>Choose your movie, and book the seat now</Typography>
-                <Grid container spacing={2}>
-                    <Grid xs={4}>
-                        <Card sx={{ maxWidth: 345 }}>
-                            <CardActionArea>
-                                <CardMedia
-                                component="img"
-                                height="140"
-                                image="/static/images/cards/contemplative-reptile.jpg"
-                                alt="green iguana"
-                                />
-                                <CardContent>
-                                <Typography gutterBottom variant="h5" component="div">
-                                    Lizard
-                                </Typography>
-                                <Typography variant="body2" color="text.secondary">
-                                    Lizards are a widespread group of squamate reptiles, with over 6,000
-                                    species, ranging across all continents except Antarctica
-                                </Typography>
-                                </CardContent>
-                            </CardActionArea>
-                            <CardActions>
-                                <Button size="small" color="primary">
-                                Share
-                                </Button>
-                            </CardActions>
-                        </Card>
-                    </Grid>
-                    <Grid xs={4}>
-                        <Card sx={{ maxWidth: 345 }}>
-                            <CardActionArea>
-                                <CardMedia
-                                component="img"
-                                height="140"
-                                image="/static/images/cards/contemplative-reptile.jpg"
-                                alt="green iguana"
-                                />
-                                <CardContent>
-                                <Typography gutterBottom variant="h5" component="div">
-                                    Lizard
-                                </Typography>
-                                <Typography variant="body2" color="text.secondary">
-                                    Lizards are a widespread group of squamate reptiles, with over 6,000
-                                    species, ranging across all continents except Antarctica
-                                </Typography>
-                                </CardContent>
-                            </CardActionArea>
-                            <CardActions>
-                                <Button size="small" color="primary">
-                                Share
-                                </Button>
-                            </CardActions>
-                        </Card>
-                    </Grid>
-                    <Grid xs={4}>
-                        <Card sx={{ maxWidth: 345 }}>
-                            <CardActionArea>
-                                <CardMedia
-                                component="img"
-                                height="140"
-                                image="/static/images/cards/contemplative-reptile.jpg"
-                                alt="green iguana"
-                                />
-                                <CardContent>
-                                <Typography gutterBottom variant="h5" component="div">
-                                    Lizard
-                                </Typography>
-                                <Typography variant="body2" color="text.secondary">
-                                    Lizards are a widespread group of squamate reptiles, with over 6,000
-                                    species, ranging across all continents except Antarctica
-                                </Typography>
-                                </CardContent>
-                            </CardActionArea>
-                            <CardActions>
-                                <Button size="small" color="primary">
-                                Share
-                                </Button>
-                            </CardActions>
-                        </Card>
-                    </Grid>
-                </Grid>
-            </div>
-        </container>
-    );
-}
-
-export default Main;
-
-if (document.getElementById('app')) {
-    ReactDOM.render(<Main />, document.getElementById('app'));
-}
